removed clippy warnings zorz42 refused to fix

+++ libraries/graphics/renderer.rs
@@ -304,12 +304,12 @@ impl Renderer {
                 gl::BlitFramebuffer(
                     0,
                     0,
-                    (self.get_window_size().0 as f32 * 2.0) as i32,
-                    (self.get_window_size().1 as f32 * 2.0) as i32,
+                    (self.get_window_size().0 * 2.0) as i32,
+                    (self.get_window_size().1 * 2.0) as i32,
                     0,
                     0,
-                    (self.get_window_size().0 as f32 * 2.0) as i32,
-                    (self.get_window_size().1 as f32 * 2.0) as i32,
+                    (self.get_window_size().0 * 2.0) as i32,
+                    (self.get_window_size().1 * 2.0) as i32,
                     gl::COLOR_BUFFER_BIT,
                     gl::NEAREST,
                 );
